font-install: Refactor Font class to its own file
